fix(model): test now only targets remote if model not provided
test allows specifying model-file
