+++ example/model.fga
@@ -6,7 +6,7 @@ type folder
     define parent: [folder]
     define owner: [user]
     define parent_owner: owner from parent or parent_owner from parent
-    define viewer: [user] or owner
+    define viewer: [user] or owner or parent_owner
     define can_share: owner
     define can_write: owner or parent_owner
-    define can_view: viewer
+    define can_view: viewer or viewer from parent
